kube subpackage: resuable mount-related routines, rewritten existing apis, started unit-test-api

+++ repackage-pyace.ps1
@@ -1,9 +1,12 @@
 Set-Location ".\pyace\"
 
 foreach($folder in "build","dist","pyace.egg-info"){
-	Remove-Item -path $folder -recurse
+	If (Test-Path $folder){
+		Remove-Item -path $folder -recurse
+	}
 }
 
+# MAKE SURE TO 'pip install wheel' FIRST
 python setup.py sdist bdist_wheel
 pip install ./dist/pyace-0.0.1.tar.gz
 Set-Location ".."
